53 frontend add api integration for activities

+++ dra2/src/main/java/com/tab/dra2/dto/ActivityTypeResponseDto.java
@@ -3,9 +3,11 @@ package com.tab.dra2.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.lang.Long;
+
 @Data
 @Builder
 public class ActivityTypeResponseDto {
-    private int id;
+    private Long id;
     private String actType;
 }
